changed directory structure in test folders accordingly
