+++ src/test/scala/de/htwg/se/ludo/model/gameComponent/gameBaseImpl/CellSpec.scala
@@ -1,8 +1,7 @@
-package de.htwg.se.ludo.model
+package de.htwg.se.ludo.model.gameComponent.gameBaseImpl
 
-import de.htwg.se.ludo.model.gameComponent.gameBaseImpl.Cell
-import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
 
 class CellSpec extends AnyWordSpec with Matchers{
   "A Cell" when { "not set to any value " should{
