+++ src/test/scala/de/htwg/se/ludo/model/gameComponent/gameBaseImpl/GameSpec.scala
@@ -1,9 +1,8 @@
-package de.htwg.se.ludo.model
+package de.htwg.se.ludo.model.gameComponent.gameBaseImpl
 
-import de.htwg.se.ludo.model.gameComponent.gameBaseImpl._
 import de.htwg.se.ludo.model.playerComponent.{Player, Team}
-import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
 
 import java.awt.Color
